% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frescalo.r
\name{benchmark_proportions}
\alias{benchmark_proportions}
\title{Extract the proportion of benchmark species found for each site and time.
These are used for effort correction when computing time factors.}
\usage{
benchmark_proportions(object)
}
\arguments{
\item{object}{An object as returned from the frescalo function.}
}
\value{
A dataframe with the proportion of benchmark species found.
}
\description{
Extract the proportion of benchmark species found for each site and time.
These are used for effort correction when computing time factors.
}
\examples{
data(bryophyte)
weights = compute_weights(hectad_locations, vascular_plants)
fr = frescalo(bryophyte, weights)
head(benchmark_proportions(fr))
}
