% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frescalo.r
\name{benchmark_species}
\alias{benchmark_species}
\title{Extract the species that are considered benchmarks for each site.}
\usage{
benchmark_species(object)
}
\arguments{
\item{object}{An object as returned from the frescalo function.}
}
\value{
A dataframe with the species used as benchmarks for each site.
}
\description{
Extract the species that are considered benchmarks for each site.
}
\examples{
data(bryophyte)
weights = compute_weights(hectad_locations, vascular_plants)
fr = frescalo(bryophyte, weights)
head(benchmark_species(fr))
}
