% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{bryophyte}
\alias{bryophyte}
\alias{hectad_locations}
\alias{vascular_plants}
\title{Bryophyte data}
\format{
\subsection{\code{bryophyte}}{

A data frame with 78332 rows and 3 columns:
\describe{
\item{hectad}{Id of hectad (site) where species was observed.}
\item{species}{Id of species.}
\item{year}{Year of observation.}
}
}

\subsection{\code{hectad_locations}}{

A data frame with 404 rows and 3 columns:
\describe{
\item{hectad}{Id of target hectad.}
\item{x}{x coordinate of hectad.}
\item{y}{y coordinate of hectad.}
}
}

\subsection{\code{vascular_plants}}{

A data frame with 404 rows and 1192 columns:
\describe{
\item{hectad}{Id of target hectad.}
\item{other columns}{Presence of absence of vascular plant species in hectad. One column for each vascular plant species.}
}
}
}
\source{
\url{https://www.brc.ac.uk/biblio/frescalo-computer-program-analyse-your-biological-records}
}
\usage{
bryophyte

hectad_locations

vascular_plants
}
\description{
A subset of bryophyte observations analyzed in Hill 2012, and
distributed with the fortran frescalo code.
}
\references{
Hill, M. O. 2012. Local frequency as a key to interpreting species occurrence data when recording effort is not known.
}
\keyword{datasets}
