% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.r
\name{compute_weights}
\alias{compute_weights}
\title{Compute neighbourhood weights according to Hill 2012.}
\usage{
compute_weights(coords, attributes, max_sp = 200, max_neigh = 100, cols = NULL)
}
\arguments{
\item{coords}{Data frame specifying locations of sites. Should have three columns:
a column with site labels, a columns with x coordinates and a columns with y coordinates.}

\item{attributes}{Data frame containing habitat or community data. Should have a column with site labels
and columns with habitat or comunity composition.}

\item{max_sp}{Maximum number of spatial neighbours to consider.}

\item{max_neigh}{Maximum neighbourhood size. Must be smaller than max_sp.}

\item{cols}{A list or vector with elements named site, x, and y and values giving the corresponding
column names in coords and attributes. If NULL, the order of the columns in coords and attributes
are used to identify the content.}
}
\value{
A data frame with columns site, neigh, and the corresponding weight.
}
\description{
Compute neighbourhood weights according to Hill 2012.
}
\examples{
weights = compute_weights(hectad_locations, vascular_plants)
}
