% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.r
\name{euclid_dist}
\alias{euclid_dist}
\title{Compute euclidean distances and their ranks coordinates.}
\usage{
euclid_dist(crd, labels = NULL, max_neigh = 200, max_dist = Inf)
}
\arguments{
\item{crd}{A data frame with x and y coordinates.}

\item{labels}{Vector with names of sites. Should have length equal to number of columns in crd.}

\item{max_neigh}{The largest number of neighbours to keep. Defaults to keeping the 200 closest neighbours of each site.}

\item{max_dist}{Distances larger than this are omitted in the output. Not applied by default.}
}
\value{
A data frame with distances.

The function computes simple euclidian distances. For more accurate results,
use a dedicated function, such as st_dist from the sf package.
}
\description{
Compute euclidean distances and their ranks coordinates.
}
\examples{
euclid_dist(hectad_locations[,c("x", "y")], labels = hectad_locations$hectad)
}
