% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frescalo.r
\name{occupancy_prob}
\alias{occupancy_prob}
\title{Compute estimated occupancy probabilities for a given level of effort.}
\usage{
occupancy_prob(object, species, s = 1)
}
\arguments{
\item{object}{An object as returned from the frescalo function.}

\item{species}{A character vector of species names for which to compute probabilities.}

\item{s}{Level of effort.}
}
\description{
Compute estimated occupancy probabilities for a given level of effort.
}
\note{
Estimated probability of occurence under standard effort, sit = 1 meaning all benchmarks found (Bijlsma).
This depends on the proportion of benchmarks (see Prescott 2025).
Note: this is rather probability of detection under a sampling effort sufficient for all benchmarks to be found(?)
}
