% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frescalo.r
\name{recording_probs}
\alias{recording_probs}
\title{Compute estimated recording probabilities for a given level of effort.}
\usage{
recording_probs(object, species, s = 1)
}
\arguments{
\item{object}{An object as returned from the frescalo function.}

\item{species}{A character vector of species names for which to compute probabilities.}

\item{s}{Level of effort.}
}
\value{
A data.frame with estimated occupancy probabilities.
}
\description{
Compute estimated recording probabilities for a given level of effort.
}
\note{
Estimated probability of recording species under standard effort, sit = 1 meaning all benchmarks found.
This depends on the proportion of benchmarks.
}
