% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frescalo.r
\name{sites}
\alias{sites}
\title{Extract information about sites from a frescalo object.}
\usage{
sites(object)
}
\arguments{
\item{object}{An object as returned from the frescalo function.}
}
\value{
A data frame with a row for each sites, and information about its neighbourhood.

\tabular{llll}{
\strong{column} \tab \strong{fortran} \tab \strong{meaning}\cr
\code{site} \tab \code{Location} \tab Name of site.\cr
\code{species_count} \tab \code{No_spp} \tab Number of species observed at site.\cr
\code{phi_obs} \tab \code{Phi_in} \tab Values of phi (frequency-weighted mean frequency) computed from observed frequencies.\cr
\code{alpha} \tab \code{Alpha} \tab Sampling effort multiplier required to reach \code{phi_target}.\cr
\code{spnum_obs} \tab \code{Spnum_in} \tab Sum of observed neighbourhood frequencies, a raw estimate of number of species. \cr
\code{spnum_est} \tab \code{Spnum_out} \tab Sum of neighbourhood frequencies after rescaling, an estimate of number of species. \cr
\code{wgt_n2} \tab \code{Wgt_n2} \tab Ratio of square of sum of neighbourhood weights to the sum of squared neighbourhood weights.
Higher values indicate higher similarity among sites in neighbourhood. \cr
\code{phi_out} \tab \code{Phi_out} \tab Values of phi after rescaling. Should be close to \code{phi_target} if algorithm converged. \cr
\code{iter_fresca} \tab \code{Iter} \tab Number of iterations required for the rescaling routine. \cr
\code{conv_fresca} \tab  \tab Indicates whether the rescaling routine converged.\cr
}
}
\description{
Extract information about sites from a frescalo object.
}
\examples{
weights = compute_weights(hectad_locations, vascular_plants)
fr = frescalo(bryophyte, weights)
head(sites(fr))
}
