% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frescalo.r
\name{summary.frescalo}
\alias{summary.frescalo}
\alias{print.summary.frescalo}
\title{Summarize frescalo output}
\usage{
\method{summary}{frescalo}(object, ...)

\method{print}{summary.frescalo}(x, ...)
}
\arguments{
\item{object}{A frescalo object.}

\item{...}{Other arguments.}

\item{x}{An object returned by \code{summary.frescalo}.}
}
\description{
Summarize frescalo output
}
