% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frescalo.r
\name{timefactors}
\alias{timefactors}
\title{Extract time factors from a frescalo object.}
\usage{
timefactors(object)
}
\arguments{
\item{object}{An object as returned from the frescalo function.}
}
\value{
A data frame with time factors across species.
}
\description{
Extract time factors from a frescalo object.
}
\examples{
weights = compute_weights(hectad_locations, vascular_plants)
fr = frescalo(bryophyte, weights)
head(timefactors(fr))
}
