% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frescalo.r
\name{check_rescaling}
\alias{check_rescaling}
\title{Plot unscaled and rescaled neighbourhood species frequency curves.}
\usage{
check_rescaling(object, max_sites = 500)
}
\arguments{
\item{object}{An object as returned from the frescalo function.}

\item{max_sites}{Maximum number of sites for which to plot curves.
If less than the total number of sites, a random sample is taken.}
}
\value{
A ggplot object.

The plots correspond to Fig. 2 and 3 in Hill 2012, a comparison of neighborhoud frequency curves
before and after rescaling.
}
\description{
Plot unscaled and rescaled neighbourhood species frequency curves.
}
\examples{
data(bryophyte)
fr = frescalo(bryophyte, bryophyte_weights)
check_rescaling(fr, max_sites = 50)
}
