% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.r
\name{dist2df.mat}
\alias{dist2df.mat}
\title{Convenience function for converting a distance matrix, generated e.g. by dist or vegan::dist
to a data.frame suitable for the frescalo function.}
\usage{
dist2df.mat(D, ids = seq_len(nrow(D)))
}
\arguments{
\item{D}{Distance matrix}

\item{ids}{Labels or ids for the sites.}
}
\value{
description A data frame with distances between pairs of sites.
}
\description{
Convenience function for converting a distance matrix, generated e.g. by dist or vegan::dist
to a data.frame suitable for the frescalo function.
}
