% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.r
\name{euclid_dist}
\alias{euclid_dist}
\title{Compute euclidean distances and their ranks coordinates.}
\usage{
euclid_dist(data, site, x, y, max_neigh = 200, max_dist = Inf)
}
\arguments{
\item{data}{A data frame with site names and x and y coordinates.}

\item{site}{Name of column with site/location name or id.}

\item{x}{Name of column containing x coordinate.}

\item{y}{Name of column containing y coordinate.}

\item{max_neigh}{The largest number of neighbours to keep. Defaults to keeping the 200 closest neighbours of each site.}

\item{max_dist}{Distances larger than this are omitted in the output. Not applied by default.}
}
\value{
A data frame with distances.

The function computes simple euclidian distances. For more accurate results,
use a dedicated function, such as st_dist from the sf package.
}
\description{
Compute euclidean distances and their ranks coordinates.
}
