% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frescalo.r
\name{frequencies}
\alias{frequencies}
\title{Extract species frequencies from a frescalo object.}
\usage{
frequencies(object)
}
\arguments{
\item{object}{An object as returned from the frescalo function.}
}
\value{
A data frame with species frequencies across locations.
}
\description{
Extract species frequencies from a frescalo object.
}
\examples{
data(bryophyte)
fr = frescalo(bryophyte, bryophyte_weights)
head(frequencies(fr))
}
