% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frescalo.r
\name{frescalo}
\alias{frescalo}
\title{Analyse species occurrence data with the frescalo algorithm of Hill 2012.}
\usage{
frescalo(
  data,
  weights,
  phi_target = 0.74,
  Rstar = 0.27,
  phi_prob = 0.985,
  bench_exclude = NULL,
  data_names = NULL,
  weights_names = NULL
)
}
\arguments{
\item{data}{Data frame with the samples. By default, first column is interpreted as the name or id of the sampled site,
the second column as the observed species, and the third column the time of observation. This can be changed
via the \code{data_names} and \code{weihgts_names} arguments.}

\item{weights}{Data frame with neighborhood weights where the first column is the target site, second column is the
neighbor, and third column is the weight in the neighborhood of the target site.}

\item{phi_target}{Target value for adjusted frequency weighted mean frequencies. The default value, 0.74, follows the default of Hill,
but is arbitrary. If NA, target is set to the quantile of the input mean frequencies corresponding to \code{phi_prob}.}

\item{Rstar}{Threshold for species to be considered as benchmarks when computing time factors.}

\item{phi_prob}{Used to check that \code{phi_target} is not too low. A warning is generated if thee quantile of input
mean frequencies corresponding to \code{phi_prob} is larger than \code{phi_target}.
If \code{phi_target} is set to NA, the quantile corresponding to \code{phi_prob} is taken as the target.
Defaults to 0.985.}

\item{bench_exclude}{Vector of names of species not to be used as benchmarks when computing time factors.}

\item{data_names}{A named list or character vector with elements named location, spec, time,
and values equal to the corresponding column names in data.
Defaults to NULL in which case the order of the columns is used.}

\item{weights_names}{A named list or character vector with elements named location, neigh, weight,
and values equal to the corresponding column names in data.
Defaults to NULL in which case the order of the columns is used.}
}
\value{
A frescalo object.
}
\description{
Analyse species occurrence data with the frescalo algorithm of Hill 2012.
}
\details{
The implementation uses similar conventions to the original fortran program. E.g. small constants are added in strategic places
to avoid divisions by zero or other issues that can cause the algorithm to otherwise fail numerically.
}
\examples{
data(bryophyte)
fr = frescalo(bryophyte, bryophyte_weights)
summary(fr)
head(frequencies(fr))
head(timefactors(fr))
}
