% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frescaloDT-package.R
\docType{package}
\name{frescaloDT-package}
\alias{frescaloDT}
\alias{frescaloDT-package}
\title{frescaloDT: Calculate Species Trends and Distribution from Presence Data with the Frescalo Algorithm}
\description{
Implements the frescalo algorithm of Hill (2012) for estimating species distributions from species presences. Uses data.table under the hood for efficiency with large data sets.
}
\author{
\strong{Maintainer}: Jonas Knape \email{jonas.knape@slu.se} (\href{https://orcid.org/0000-0002-8012-5131}{ORCID})

}
\keyword{internal}
